Move duplicated logic into OD test helpers trait
