+++ plugins/auto-sizes/tests/test-optimization-detective.php
@@ -5,7 +5,11 @@
  * @package auto-sizes
  */
 
+require_once __DIR__ . '/../../../tests/class-optimization-detective-test-helpers.php';
+
 class Test_Auto_Sizes_Optimization_Detective extends WP_UnitTestCase {
+	use Optimization_Detective_Test_Helpers;
+
 	/**
 	 * Runs the routine before each test is executed.
 	 */
@@ -102,25 +106,7 @@ class Test_Auto_Sizes_Optimization_Detective extends WP_UnitTestCase {
 	 * @phpstan-param array<string, mixed> $element_metrics
 	 */
 	public function test_od_optimize_template_output_buffer( array $element_metrics, string $buffer, string $expected ): void {
-		$slug        = od_get_url_metrics_slug( od_get_normalized_query_vars() );
-		$sample_size = od_get_url_metrics_breakpoint_sample_size();
-		foreach ( array_merge( od_get_breakpoint_max_widths(), array( 1000 ) ) as $viewport_width ) {
-			for ( $i = 0; $i < $sample_size; $i++ ) {
-				OD_URL_Metrics_Post_Type::store_url_metric(
-					$slug,
-					$this->get_validated_url_metric(
-						$viewport_width,
-						array(
-							$element_metrics,
-						)
-					)
-				);
-			}
-		}
-
-		$remove_initial_tabs = static function ( string $input ): string {
-			return (string) preg_replace( '/^\t+/m', '', $input );
-		};
+		$this->populate_url_metrics( array( $element_metrics ) );
 
 		$html_start_doc = '<html lang="en"><head><meta charset="utf-8"><title>...</title></head><body>';
 		$html_end_doc   = '</body></html>';
@@ -130,49 +116,9 @@ class Test_Auto_Sizes_Optimization_Detective extends WP_UnitTestCase {
 		$buffer = preg_replace( '#</body>.*$#s', '', $buffer );
 
 		$this->assertEquals(
-			$remove_initial_tabs( $expected ),
-			$remove_initial_tabs( $buffer ),
+			$this->remove_initial_tabs( $expected ),
+			$this->remove_initial_tabs( $buffer ),
 			"Buffer snapshot:\n$buffer"
 		);
-	}
-
-	/**
-	 * Gets a validated URL metric.
-	 *
-	 * @param int                                      $viewport_width Viewport width for the URL metric.
-	 * @param array<array{xpath: string, isLCP: bool}> $elements       Elements.
-	 * @return OD_URL_Metric URL metric.
-	 * @throws Exception From OD_URL_Metric if there is a parse error, but there won't be.
-	 */
-	private function get_validated_url_metric( int $viewport_width, array $elements = array() ): OD_URL_Metric {
-		$data = array(
-			'url'       => home_url( '/' ),
-			'viewport'  => array(
-				'width'  => $viewport_width,
-				'height' => 800,
-			),
-			'timestamp' => microtime( true ),
-			'elements'  => array_map(
-				static function ( array $element ): array {
-					return array_merge(
-						array(
-							'isLCPCandidate'     => true,
-							'intersectionRatio'  => 1,
-							'intersectionRect'   => array(
-								'width'  => 100,
-								'height' => 100,
-							),
-							'boundingClientRect' => array(
-								'width'  => 100,
-								'height' => 100,
-							),
-						),
-						$element
-					);
-				},
-				$elements
-			),
-		);
-		return new OD_URL_Metric( $data );
 	}
 }
