+++ plugins/embed-optimizer/tests/test-optimization-detective.php
@@ -5,10 +5,11 @@
  * @package embed-optimizer
  */
 
-/**
- * @phpstan-type ElementDataSubset array{xpath: string, isLCP: bool, intersectionRatio: float}
- */
+require_once __DIR__ . '/../../../tests/class-optimization-detective-test-helpers.php';
+
 class Test_Embed_Optimizer_Optimization_Detective extends WP_UnitTestCase {
+	use Optimization_Detective_Test_Helpers;
+
 	/**
 	 * Runs the routine before each test is executed.
 	 */
@@ -57,10 +58,6 @@ class Test_Embed_Optimizer_Optimization_Detective extends WP_UnitTestCase {
 	public function test_od_optimize_template_output_buffer( Closure $set_up, string $buffer, string $expected ): void {
 		$set_up( $this );
 
-		$remove_initial_tabs = static function ( string $input ): string {
-			return (string) preg_replace( '/^\t+/m', '', $input );
-		};
-
 		$buffer = od_optimize_template_output_buffer( $buffer );
 		$buffer = preg_replace_callback(
 			':(<script type="module">)(.+?)(</script>):s',
@@ -76,73 +73,9 @@ class Test_Embed_Optimizer_Optimization_Detective extends WP_UnitTestCase {
 			$buffer
 		);
 		$this->assertEquals(
-			$remove_initial_tabs( $expected ),
-			$remove_initial_tabs( $buffer ),
+			$this->remove_initial_tabs( $expected ),
+			$this->remove_initial_tabs( $buffer ),
 			"Buffer snapshot:\n$buffer"
 		);
-	}
-
-	/**
-	 * Populates complete URL metrics for the provided element data.
-	 *
-	 * @phpstan-param ElementDataSubset[] $elements
-	 * @param array[] $elements Element data.
-	 * @param bool    $complete Whether to fully populate the groups.
-	 * @throws Exception But it won't.
-	 */
-	public function populate_url_metrics( array $elements, bool $complete = true ): void {
-		$slug        = od_get_url_metrics_slug( od_get_normalized_query_vars() );
-		$sample_size = $complete ? od_get_url_metrics_breakpoint_sample_size() : 1;
-		foreach ( array_merge( od_get_breakpoint_max_widths(), array( 1000 ) ) as $viewport_width ) {
-			for ( $i = 0; $i < $sample_size; $i++ ) {
-				OD_URL_Metrics_Post_Type::store_url_metric(
-					$slug,
-					$this->get_validated_url_metric(
-						$viewport_width,
-						$elements
-					)
-				);
-			}
-		}
-	}
-
-	/**
-	 * Gets a validated URL metric.
-	 *
-	 * @param int                      $viewport_width Viewport width for the URL metric.
-	 * @param array<ElementDataSubset> $elements       Elements.
-	 * @return OD_URL_Metric URL metric.
-	 * @throws Exception From OD_URL_Metric if there is a parse error, but there won't be.
-	 */
-	public function get_validated_url_metric( int $viewport_width, array $elements = array() ): OD_URL_Metric {
-		$data = array(
-			'url'       => home_url( '/' ),
-			'viewport'  => array(
-				'width'  => $viewport_width,
-				'height' => 800,
-			),
-			'timestamp' => microtime( true ),
-			'elements'  => array_map(
-				static function ( array $element ): array {
-					return array_merge(
-						array(
-							'isLCPCandidate'     => true,
-							'intersectionRatio'  => 1,
-							'intersectionRect'   => array(
-								'width'  => 100,
-								'height' => 100,
-							),
-							'boundingClientRect' => array(
-								'width'  => 100,
-								'height' => 100,
-							),
-						),
-						$element
-					);
-				},
-				$elements
-			),
-		);
-		return new OD_URL_Metric( $data );
 	}
 }
