+++ plugins/image-prioritizer/tests/test-helper.php
@@ -5,7 +5,10 @@
  * @package image-prioritizer
  */
 
+require_once __DIR__ . '/../../../tests/class-optimization-detective-test-helpers.php';
+
 class Test_Image_Prioritizer_Helper extends WP_UnitTestCase {
+	use Optimization_Detective_Test_Helpers;
 
 	/**
 	 * Test printing the meta generator tag.
@@ -49,10 +52,6 @@ class Test_Image_Prioritizer_Helper extends WP_UnitTestCase {
 	public function test_image_prioritizer_register_tag_visitors( Closure $set_up, string $buffer, string $expected ): void {
 		$set_up( $this );
 
-		$remove_initial_tabs = static function ( string $input ): string {
-			return (string) preg_replace( '/^\t+/m', '', $input );
-		};
-
 		$buffer = preg_replace(
 			':<script type="module">.+?</script>:s',
 			'<script type="module">/* import detect ... */</script>',
@@ -60,51 +59,9 @@ class Test_Image_Prioritizer_Helper extends WP_UnitTestCase {
 		);
 
 		$this->assertEquals(
-			$remove_initial_tabs( $expected ),
-			$remove_initial_tabs( $buffer ),
+			$this->remove_initial_tabs( $expected ),
+			$this->remove_initial_tabs( $buffer ),
 			"Buffer snapshot:\n$buffer"
 		);
-	}
-
-	/**
-	 * Gets a validated URL metric.
-	 *
-	 * @todo Move this into a trait or an Optimization Detective helper base test class.
-	 *
-	 * @param int                                      $viewport_width Viewport width for the URL metric.
-	 * @param array<array{xpath: string, isLCP: bool}> $elements       Elements.
-	 * @return OD_URL_Metric URL metric.
-	 * @throws Exception From OD_URL_Metric if there is a parse error, but there won't be.
-	 */
-	public function get_validated_url_metric( int $viewport_width, array $elements = array() ): OD_URL_Metric {
-		$data = array(
-			'url'       => home_url( '/' ),
-			'viewport'  => array(
-				'width'  => $viewport_width,
-				'height' => 800,
-			),
-			'timestamp' => microtime( true ),
-			'elements'  => array_map(
-				static function ( array $element ): array {
-					return array_merge(
-						array(
-							'isLCPCandidate'     => true,
-							'intersectionRatio'  => 1,
-							'intersectionRect'   => array(
-								'width'  => 100,
-								'height' => 100,
-							),
-							'boundingClientRect' => array(
-								'width'  => 100,
-								'height' => 100,
-							),
-						),
-						$element
-					);
-				},
-				$elements
-			),
-		);
-		return new OD_URL_Metric( $data );
 	}
 }
